+++ openadapt/privacy/providers/__init__.py
@@ -2,3 +2,10 @@
 
 Module: __init__.py
 """
+
+
+class ScrubProvider:  # pylint: disable=too-few-public-methods
+    """A Class for Scrubbing Provider."""
+
+    PRESIDIO = "PRESIDIO"
+    COMPREHEND = "COMPREHEND"
