+++ tests/openadapt/privacy/providers/test_comprehend_scrub.py
@@ -1,31 +1,9 @@
-"""Module to test PresidioScrubbingProvider."""
+"""Module to test ComprehendScrubbingProvider."""
 
 
-from io import BytesIO
-import os
+from openadapt.privacy.providers.aws_comprehend import ComprehendScrubbingProvider
 
-from PIL import Image
-import pytest
-import spacy
-
-from openadapt import config
-
-if not spacy.util.is_package(config.SPACY_MODEL_NAME):  # pylint: disable=no-member
-    pytestmark = pytest.mark.skip(reason="SpaCy model not installed!")
-else:
-    from openadapt.privacy.providers.presidio import PresidioScrubbingProvider
-
-    scrub = PresidioScrubbingProvider()
-
-
-def test_presidio_scrub_text() -> None:
-    """Test that PresidioScrubbingProvider can scrub text."""
-    text = "My phone number is 123-456-7890."
-    expected_result = "My phone number is <PHONE_NUMBER>."
-
-    scrubbed_text = scrub.scrub_text(text)
-
-    assert scrubbed_text == expected_result
+scrub = ComprehendScrubbingProvider()
 
 
 def _hex_to_rgb(hex_color: int) -> tuple[int, int, int]:
@@ -42,45 +20,6 @@ def _hex_to_rgb(hex_color: int) -> tuple[int, int, int]:
     green = (hex_color >> 8) & 0xFF
     red = hex_color & 0xFF
     return red, green, blue
-
-
-def test_scrub_image() -> None:
-    """Test that the scrubbed image data is different."""
-    # Read test image data from file
-    test_image_path = "assets/test_scrub_image.png"
-    with open(test_image_path, "rb") as file:
-        test_image_data = file.read()
-
-    # Convert image data to PIL Image object
-    test_image = Image.open(BytesIO(test_image_data))
-
-    # Scrub the image
-    scrubbed_image = scrub.scrub_image(test_image)
-
-    # Save the scrubbed image data to a file
-    scrubbed_image_path = "scrubbed_image.png"
-    scrubbed_image.save(scrubbed_image_path)
-
-    # Load the scrubbed image from file for manual verification
-    scrubbed_image = Image.open(scrubbed_image_path)
-    scrubbed_image = scrubbed_image.convert("RGB")
-
-    # Count the number of pixels having the color of the mask
-    mask_pixels = sum(
-        1
-        for pixel in scrubbed_image.getdata()
-        if pixel == _hex_to_rgb(config.SCRUB_FILL_COLOR)  # pylint: disable=no-member
-    )
-    total_pixels = scrubbed_image.width * scrubbed_image.height
-
-    test_image.close()
-    scrubbed_image.close()
-    os.remove(scrubbed_image_path)
-
-    # Assert ~1.5% mask pixels compared to total pixels.
-    assert (
-        round((mask_pixels / total_pixels), 3) == 0.022
-    )  # Change this value as necessary
 
 
 def test_empty_string() -> None:
@@ -100,8 +39,7 @@ def test_no_scrub_string() -> None:
 def test_scrub_email() -> None:
     """Test that the email address is scrubbed."""
     assert (
-        scrub.scrub_text("My email is john.doe@example.com.")
-        == "My email is <EMAIL_ADDRESS>."
+        scrub.scrub_text("My email is john.doe@example.com.") == "My email is <EMAIL>."
     )
 
 
@@ -109,7 +47,7 @@ def test_scrub_phone_number() -> None:
     """Test that the phone number is scrubbed."""
     assert (
         scrub.scrub_text("My phone number is 123-456-7890.")
-        == "My phone number is <PHONE_NUMBER>."
+        == "My phone number is <PHONE>."
     )
 
 
@@ -117,7 +55,7 @@ def test_scrub_credit_card() -> None:
     """Test that the credit card number is scrubbed."""
     assert (
         scrub.scrub_text("My credit card number is 4234-5678-9012-3456 and ")
-    ) == "My credit card number is <CREDIT_CARD> and "
+    ) == "My credit card number is <CREDIT_DEBIT_NUMBER> and "
 
 
 def test_scrub_date_of_birth() -> None:
@@ -132,7 +70,7 @@ def test_scrub_address() -> None:
     """Test that the address is scrubbed."""
     assert (
         scrub.scrub_text("My address is 123 Main St, Toronto, On, CAN.")
-        == "My address is 123 Main St, <LOCATION>, <LOCATION>, <LOCATION>."
+        == "My address is <ADDRESS>."
     )
 
 
@@ -141,7 +79,7 @@ def test_scrub_ssn() -> None:
     # Test scrubbing of social security number
     assert (
         scrub.scrub_text("My social security number is 923-45-6789")
-        == "My social security number is <US_SSN>"
+        == "My social security number is <SSN>"
     )
 
 
@@ -149,7 +87,7 @@ def test_scrub_dl() -> None:
     """Test that the driver's license number is scrubbed."""
     assert (
         scrub.scrub_text("My driver's license number is A123-456-789-012")
-        == "My driver's license number is <US_DRIVER_LICENSE>-456-789-012"
+        == "My driver's license number is <DRIVER_ID>"
     )
 
 
@@ -157,7 +95,7 @@ def test_scrub_passport() -> None:
     """Test that the passport number is scrubbed."""
     assert (
         scrub.scrub_text("My passport number is A1234567.")
-        == "My passport number is <US_DRIVER_LICENSE>."
+        == "My passport number is <PASSPORT_NUMBER>."
     )
 
 
@@ -165,7 +103,7 @@ def test_scrub_national_id() -> None:
     """Test that the national ID number is scrubbed."""
     assert (
         scrub.scrub_text("My national ID number is 1234567890123.")
-        == "My national ID number is <US_BANK_NUMBER>."
+        == "My national ID number is <PHONE>."
     )
 
 
@@ -173,9 +111,7 @@ def test_scrub_routing_number() -> None:
     """Test that the bank routing number is scrubbed."""
     assert (
         scrub.scrub_text("My bank routing number is 123456789.")
-        == "My bank routing number is <US_PASSPORT>."
-        or scrub.scrub_text("My bank routing number is 123456789.")
-        == "My bank routing number is <US_BANK_NUMBER>."
+        == "My bank routing number is <BANK_ROUTING>."
     )
 
 
@@ -183,7 +119,7 @@ def test_scrub_bank_account() -> None:
     """Test that the bank account number is scrubbed."""
     assert (
         scrub.scrub_text("My bank account number is 635526789012.")
-        == "My bank account number is <US_BANK_NUMBER>."
+        == "My bank account number is <BANK_ACCOUNT_NUMBER>."
     )
 
 
@@ -198,10 +134,10 @@ def test_scrub_all_together() -> None:
     )
     assert (
         scrub.scrub_text(text_with_pii_phi)
-        == "<PERSON> email is <EMAIL_ADDRESS> and"
-        " his phone number is <PHONE_NUMBER>."
-        "His credit card number is <CREDIT_CARD> and"
-        " his social security number is <US_SSN>."
+        == "<NAME>'s email is <EMAIL> and"
+        " his phone number is <PHONE>"
+        " credit card number is <CREDIT_DEBIT_NUMBER> and"
+        " his social security number is <SSN>."
         " He was born on <DATE_TIME>."
     )
 
@@ -210,6 +146,6 @@ def test_scrub_dict() -> None:
     """Test that the scrub_dict function works."""
     text_with_pii_phi = {"title": "hi my name is Bob Smith."}
 
-    expected_output = {"title": "hi my name is <PERSON>."}
+    expected_output = {"title": "hi my name is <NAME>."}
 
     assert scrub.scrub_dict(text_with_pii_phi) == expected_output
